Initial reshuffle of vumi.application.sandbox.
