+++ vumi/application/sandbox/__init__.py
@@ -1,0 +1,14 @@
+"""A worker for executing code in a sandbox."""
+
+__all__ = [
+    "JsSandbox", "JsFileSandbox", "SandboxResource",
+    "LoggingResource", "HttpClientResource", "OutboundResource",
+    "RedisResource",
+]
+
+from .worker import JsSandbox, JsFileSandbox
+from .resources.utils import SandboxResource
+from .resources.logging import LoggingResource
+from .resources.http import HttpClientResource
+from .resources.kv import RedisResource
+from .resources.outbound import OutboundResource
