+++ vumi/application/sandbox/tests/app_requires_path.js
