+++ vumi/application/sandbox/utils.py
@@ -1,0 +1,2 @@
+class SandboxError(Exception):
+    """An error occurred inside the sandbox."""
